% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojizer.R
\name{emojizer}
\alias{emojizer}
\title{Optimizer: Emojizer}
\usage{
emojizer(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs one emojizer pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "train <- function(param) {",
  "  param + 100",
  "}",
  sep = "\\n"
)
cat(emojizer(list(code))$codes[[1]])
}
