% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_to_lapply.R
\name{for_to_lapply}
\alias{for_to_lapply}
\title{Convert 'for' loop code to 'lapply' code}
\usage{
for_to_lapply(for_text)
}
\arguments{
\item{for_text}{a character vector with the runnable 'for' loop code.}
}
\description{
Convert 'for' loop code to 'lapply' code
}
\examples{
for_to_lapply("
  for (i in 1:10) {
    # just prints the square of numbers from 1 to 10
    print(i^2)
  }
")

}
