% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-fun-files.R
\name{generate_files_opt_report}
\alias{generate_files_opt_report}
\title{Report possible optimizations in `.R` files.}
\usage{
generate_files_opt_report(files, optimizers = rco:::all_optimizers)
}
\arguments{
\item{files}{A character vector with paths to files to optimize.}

\item{optimizers}{A named list of optimizer functions.}
}
\description{
Report possible optimizations in `.R` files.
}
