% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-fun-folder.R
\name{generate_folder_opt_report}
\alias{generate_folder_opt_report}
\title{Report possible optimizations in folder containing `.R` files.}
\usage{
generate_folder_opt_report(
  folder,
  optimizers = all_optimizers,
  pattern = "\\\\.R$",
  recursive = TRUE
)
}
\arguments{
\item{folder}{Path to a directory with files to optimize.}

\item{optimizers}{A named list of optimizer functions.}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be optimized.}

\item{recursive}{A logical value indicating whether or not files
in subdirectories of `folder` should be optimized as well.}
}
\description{
Report possible optimizations in folder containing `.R` files.
}
