% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-fun-text.R
\name{generate_text_opt_report}
\alias{generate_text_opt_report}
\title{Report possible optimizations in `.R` code snippet.}
\usage{
generate_text_opt_report(text, optimizers = all_optimizers)
}
\arguments{
\item{text}{A character vector with the code to optimize.}

\item{optimizers}{A named list of optimizer functions.}
}
\description{
Report possible optimizations in `.R` code snippet.
}
