% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\docType{data}
\name{max_optimizers}
\alias{max_optimizers}
\title{Max optimizers list.}
\format{An object of class \code{list} of length 7.}
\usage{
max_optimizers
}
\description{
List of all the optimizer functions, with maximum optimization techniques
enabled.
Note that using this optimizers could change the semantics of the program!
\itemize{
  \item Constant Folding \code{\link{opt_constant_folding}}
  \item Constant Propagation \code{\link{opt_constant_propagation}}
  \item Dead Code Elimination \code{\link{opt_dead_code}}
  \item Dead Store Elimination \code{\link{opt_dead_store}}
  \item Dead Expression Elimination \code{\link{opt_dead_expr}}
  \item Common Subexpression Elimination \code{\link{opt_common_subexpr}}
  \item Loop-invariant Code Motion \code{\link{opt_loop_invariant}}
}
}
\keyword{datasets}
