% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-column-extractor.R
\name{opt_column_extractor}
\alias{opt_column_extractor}
\title{Optimizer: Efficient Column Extraction.}
\usage{
opt_column_extractor(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs optimisations for column extraction throughout the code.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
 "mtcars[ , 11]",
 "#This is a comment!!",
 "mtcars [[11]]",
 "mtcars$carb",
 "yo <- 1",
 "yo",
 "mtcars[[c(11)]]",
 ".subset2(mtcars, 11)",
  sep = "\n"
)
cat(opt_column_extractor(list(code))$codes[[1]])
}
