% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-common-subexpr.R
\name{opt_common_subexpr}
\alias{opt_common_subexpr}
\title{Optimizer: Common Subexpression Elimination}
\usage{
opt_common_subexpr(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs one common subexpression elimination pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "a <- b * c + g"
  "d = b * c * e",
  sep = "\\n"
)
cat(opt_common_subexpr(list(code))$codes[[1]])
}
