% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-cond-thread.R
\name{opt_cond_thread}
\alias{opt_cond_thread}
\title{Optimizer: Conditional Threading.}
\usage{
opt_cond_thread(code)
}
\arguments{
\item{code}{A list of character vectors with the code to optimize.}
}
\description{
Performs one conditional threading pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
"num <- sample(1:100, 1)",
"even_sum <- 0",
"odd_sum_a <- 0",
"odd_sum_b <- 0",
"if(num \%\% 2 == 1) {",
"  odd_sum_a <- odd_sum_a + num",
"}",
"if(num \%\% 2 == 1) {",
"  odd_num_b <- odd_num_b + num",
"}",
"if(!(num \%\% 2 == 1)) {",
"  even_sum <- even_sum + num",
"}",
sep = "\n"
)

cat(opt_cond_thread(list(code))$codes[[1]])
}
