% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-constant-folding.R
\name{opt_constant_folding}
\alias{opt_constant_folding}
\title{Optimizer: Constant Folding}
\usage{
opt_constant_folding(texts, fold_floats = FALSE)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}

\item{fold_floats}{Logical indicating if floating-point results should be
folded (will reduce precision).}
}
\description{
Performs one constant folding pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste("i <- 320 * 200 * 32",
  "x <- i * 20 + 100",
  sep = "\\n"
)
opt_constant_folding(list(code))
}
