% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-dead-code.R
\name{opt_dead_code}
\alias{opt_dead_code}
\title{Optimizer: Dead Code Elimination.}
\usage{
opt_dead_code(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs one dead code elimination pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "while (TRUE) {",
  "  break",
  "  dead_code()",
  "}",
  sep = "\n"
)
cat(opt_dead_code(list(code))$codes[[1]])
}
