% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-dead-expr.R
\name{opt_dead_expr}
\alias{opt_dead_expr}
\title{Optimizer: Dead Expression Elimination.}
\usage{
opt_dead_expr(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs one dead expression elimination pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "foo <- function(x) {",
  "  x ^ 3",
  "  return(x ^ 3)",
  "}",
  sep = "\n"
)
cat(opt_dead_expr(list(code))$codes[[1]])
}
