% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-dead-store.R
\name{opt_dead_store}
\alias{opt_dead_store}
\title{Optimizer: Dead Store Elimination.}
\usage{
opt_dead_store(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs one dead store elimination pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "foo <- function() {",
  "  x <- 128 ^ 2",
  "  return(TRUE)",
  "}",
  sep = "\n"
)
cat(opt_dead_store(list(code))$codes[[1]])
}
