% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-memory-alloc.R
\name{opt_memory_alloc}
\alias{opt_memory_alloc}
\title{Optimizer: Memory Allocation.}
\usage{
opt_memory_alloc(code)
}
\arguments{
\item{code}{A list of character vectors with the code to optimize.}
}
\description{
Performs one memory allocation pass.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
  "v <- NULL",
  "for (i in 1:5) {",
  "  v[i] <- i^2",
  "}",
  sep = "\n"
)
cat(opt_memory_alloc(list(code))$codes[[1]])
}
