% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-value-extractor.R
\name{opt_value_extractor}
\alias{opt_value_extractor}
\title{Optimizer: Efficient Value Extraction.}
\usage{
opt_value_extractor(texts)
}
\arguments{
\item{texts}{A list of character vectors with the code to optimize.}
}
\description{
Performs optimisations for value extraction throughout the code.
Carefully examine the results after running this function!
}
\examples{
code <- paste(
"mtcars[32, 11]",
"#This is a comment!!",
"mtcars [[11]][32]",
"mtcars$carb[32]",
"yo <- 1",
"yo",
"mtcars[[c(11,32)]]",
".subset2(mtcars, 11)[32]",
 sep = "\n")
 )
cat(opt_column_extractor(list(code))$codes[[1]])
}
