% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-text.R
\name{optimize_text}
\alias{optimize_text}
\title{Optimize text containing code}
\usage{
optimize_text(text, optimizers)
}
\arguments{
\item{text}{A character vector with the code to optimize.}

\item{optimizers}{A list of optimizer functions.}
}
\description{
Performs the desired optimization strategies in the text.
Carefully examine the results after running this function!
}
