% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco-gui.R
\name{rco_gui}
\alias{rco_gui}
\title{rco GUI selector.}
\usage{
rco_gui(option)
}
\arguments{
\item{option}{A character indicating which GUI to open.
One from:
\itemize{
  \item "code_optimizer" for single code optimizing.
  \item "pkg_optimizer" for package optimizing.
}}
}
\description{
Starts the selected rco Graphical User Interface (GUI).
}
\examples{
## Start the GUI
\donttest{
rco_gui("")
}
}
